requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
psycopg[binary]==3.2.1
bcrypt==4.2.0
PyJWT==2.9.0
bandit==1.7.9
safety==3.2.0